# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grainily API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/units:
    get:
      operationId: adminListUnits
      tags: [admin]
      summary: Admin - list all units
      responses:
        "200":
          description: All units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: adminCreateUnit
      tags: [admin]
      summary: Admin - create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /admin/units/{id}:
    put:
      operationId: adminUpdateUnit
      tags: [admin]
      summary: Admin - update unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: adminDeleteUnit
      tags: [admin]
      summary: Admin - delete unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/promotion-prices:
    get:
      operationId: adminListPromotionPrices
      tags: [admin]
      summary: Admin - list promotion prices
      responses:
        "200":
          description: All promotion prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromotionPrice"
    post:
      operationId: adminCreatePromotionPrice
      tags: [admin]
      summary: Admin - create promotion price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionPriceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionPrice"
  /admin/promotion-prices/{id}:
    put:
      operationId: adminUpdatePromotionPrice
      tags: [admin]
      summary: Admin - update promotion price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionPriceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionPrice"
    delete:
      operationId: adminDeletePromotionPrice
      tags: [admin]
      summary: Admin - delete promotion price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Unit:
      type: object
      required: [id, name, symbol, active]
      properties:
        id:
          type: integer
        name:
          type: string
        symbol:
          type: string
        active:
          type: boolean
    UnitInput:
      type: object
      required: [name, symbol]
      properties:
        name:
          type: string
        symbol:
          type: string
        active:
          type: boolean
    PromotionPrice:
      type: object
      required: [id, duration, label, price, active]
      properties:
        id:
          type: integer
        duration:
          type: integer
        label:
          type: string
        price:
          type: string
        active:
          type: boolean
    PromotionPriceInput:
      type: object
      required: [duration, label, price]
      properties:
        duration:
          type: integer
        label:
          type: string
        price:
          type: string
        active:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT